+++ persistence/src/main/java/ar/edu/itba/paw/persistence/CourseDaoJpa.java
@@ -167,8 +167,8 @@ public class CourseDaoJpa extends BasePaginationDaoImpl<Course> implements Cours
     }
 
     @Override
-    public Integer getTotalStudents(Long courseId){
-        TypedQuery<Integer> totalStudentsTypedQuery = em.createQuery("SELECT COUNT(DISTINCT enrollment.user.userId ) FROM Enrollment enrollment WHERE enrollment.course.courseId = :courseId AND enrollment.role.roleId = :roleId",Integer.class);
+    public Long getTotalStudents(Long courseId){
+        TypedQuery<Long> totalStudentsTypedQuery = em.createQuery("SELECT COUNT(DISTINCT enrollment.user.userId ) FROM Enrollment enrollment WHERE enrollment.course.courseId = :courseId AND enrollment.role.roleId = :roleId",Long.class);
         totalStudentsTypedQuery.setParameter("courseId", courseId);
         totalStudentsTypedQuery.setParameter("roleId", Roles.STUDENT.getValue());
         return totalStudentsTypedQuery.getSingleResult();
